+++ packages/yourloops/components/carbs-and-bolus/enums/carbs-and-bolus-time-range-ms.enum.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Diabeloop
+ * Copyright (c) 2023-2024, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,13 +25,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export enum CarbsAndBolusTimeRange {
-  Midnight = '0h',
-  Three = '3h',
-  Six = '6h',
-  Nine = '9h',
-  Twelve = '12h',
-  Fifteen = '15h',
-  Eighteen = '18h',
-  TwentyOne = '21h'
+export enum CarbsAndBolusTimeRangeMs {
+  Midnight = 0,         //  0 * ONE_HOUR_MS (ONE_HOUR_MS = 3600000)
+  Three = 10800000,     //  3 * ONE_HOUR_MS
+  Six = 21600000,       //  6 * ONE_HOUR_MS
+  Nine = 32400000,      //  9 * ONE_HOUR_MS
+  Twelve = 43200000,    // 12 * ONE_HOUR_MS
+  Fifteen = 54000000,   // 15 * ONE_HOUR_MS
+  Eighteen = 64800000,  // 18 * ONE_HOUR_MS
+  TwentyOne = 75600000, // 21 * ONE_HOUR_MS
 }